🚚 api: [CERTIF] move scenarioSimulatorBatchSerializer to src/certif/flash
