+++ api/src/certification/flash-certification/application/scenario-simulator-controller.js
@@ -3,7 +3,6 @@ import { Readable } from 'node:stream';
 import _ from 'lodash';
 
 import { HttpErrors } from '../../../../lib/application/http-errors.js';
-import { scenarioSimulatorBatchSerializer } from '../../../../lib/infrastructure/serializers/jsonapi/scenario-simulator-batch-serializer.js';
 import { parseCsv } from '../../../../scripts/helpers/csvHelpers.js';
 import { pickChallengeService } from '../../../evaluation/domain/services/pick-challenge-service.js';
 import { random } from '../../../shared/infrastructure/utils/random.js';
@@ -11,6 +10,7 @@ import { extractLocaleFromRequest } from '../../../shared/infrastructure/utils/r
 import { pickAnswerStatusService } from '../../shared/domain/services/pick-answer-status-service.js';
 import { FlashAssessmentSuccessRateHandler } from '../domain/models/FlashAssessmentSuccessRateHandler.js';
 import { usecases } from '../domain/usecases/index.js';
+import { scenarioSimulatorBatchSerializer } from '../infrastructure/serializers/scenario-simulator-batch-serializer.js';
 
 async function simulateFlashAssessmentScenario(
   request,
