+++ api/src/certification/flash-certification/infrastructure/serializers/scenario-simulator-batch-serializer.js
