+++ api/tests/certification/flash-certification/unit/infrastructure/serializers/scenario-simulator-batch-serializer_test.js
@@ -1,5 +1,5 @@
-import { scenarioSimulatorBatchSerializer } from '../../../../../lib/infrastructure/serializers/jsonapi/scenario-simulator-batch-serializer.js';
-import { domainBuilder, expect } from '../../../../test-helper.js';
+import { scenarioSimulatorBatchSerializer } from '../../../../../../src/certification/flash-certification/infrastructure/serializers/scenario-simulator-batch-serializer.js';
+import { domainBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | scenario-simulator-batch-serializer', function () {
   describe('#serialize', function () {
